stable version

+++ bronco/modelling/cone_construction.py
@@ -87,7 +87,7 @@ def is_point_in_cylinder(
     # Final result: inside the cylinder if within height bounds and inside the ellipse
     inside_cylinder = height_bounds_check & ellipse_checks
 
-    return inside_cylinder.astype(int)
+    return inside_cylinder
 
 
 if __name__ == "__main__":
